Correção dos codigos -> dto corrigidos ✅

+++ src/main/java/com/senai/ContaBancaria/Application/DTO/ServicoDTO.java
@@ -6,16 +6,19 @@ import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ServicoDTO(
         @Schema(description = "ID da conta bancaria", example = "1")
         String id,
+
         @NotBlank(message = "Informações da conta bancaria são obrigatórias")
         @Schema(description = "Descrição do conta bancaria", example = "A conta bancaria é do tipo poupança")
         String tipoConta,
 
         @NotNull(message = "O saldo é obrigatório")
         @DecimalMin(value = "0.0", inclusive = false, message = "saldo deve ser positivo")
-        Double saldo
+        BigDecimal saldo
 ) {
     public static ServicoDTO fromEntity(ServicoEntity s) {
         return new ServicoDTO(
